Move Owner_ID and Has_Connected_Admins into Authentication namespace.
