+++ includes/Core/Authentication/Owner_ID.php
@@ -1,14 +1,16 @@
 <?php
 /**
- * Class Google\Site_Kit\Core\Storage\Owner_ID
+ * Class Google\Site_Kit\Core\Authentication\Owner_ID
  *
- * @package   Google\Site_Kit\Core\Storage
+ * @package   Google\Site_Kit\Core\Authentication
  * @copyright 2020 Google LLC
  * @license   https://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://sitekit.withgoogle.com
  */
 
-namespace Google\Site_Kit\Core\Storage;
+namespace Google\Site_Kit\Core\Authentication;
+
+use Google\Site_Kit\Core\Storage\Setting;
 
 /**
  * Owner_ID class.
